+++ app/components/signup/signup-confirm-backup.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { observer } from 'mobx-react/native';
 import { Text, View, Image } from 'react-native';
 import { observable, when } from 'mobx';
-import TextBox from '../controls/textbox';
 import ActivityOverlay from '../controls/activity-overlay';
 import { vars } from '../../styles/styles';
 import signupState from './signup-state';
@@ -10,12 +9,15 @@ import { tx } from '../utils/translator';
 import LoginWizardPage, {
     header2, innerSmall, circleTopSmall, headingStyle2, row, container, topCircleSizeSmall
 } from '../login/login-wizard-page';
-import testLabel from '../helpers/test-label';
+import StyledTextInput from '../shared/styled-text-input';
+import icons from '../helpers/icons';
 
 const imageWelcomeSafe = require('../../assets/welcome-safe.png');
 
-const formStyle = {
-    padding: vars.spacing.medium.midi2x,
+const textDescription = {
+    paddingTop: vars.spacing.medium.midi2x,
+    paddingBottom: vars.spacing.small.maxi,
+    paddingHorizontal: vars.spacing.medium.midi2x,
     justifyContent: 'space-between'
 };
 
@@ -34,9 +36,11 @@ const titleDark = [textNormal, {
 @observer
 export default class SignupConfirmBackup extends LoginWizardPage {
     @observable confirmTextSample = tx('title_confirmTextSample');
-    @observable confirmText = '';
+    @observable confirmTextState = observable({ value: '' });
 
-    get isOk() { return this.confirmTextSample.toLowerCase() === this.confirmText.toLowerCase(); }
+    get isOk() {
+        return this.confirmTextSample.toLowerCase() === this.confirmTextState.value.toLowerCase();
+    }
 
     get icon() {
         const width = topCircleSizeSmall * 2;
@@ -51,11 +55,10 @@ export default class SignupConfirmBackup extends LoginWizardPage {
     }
 
     componentDidMount() {
-        when(() => this.isOk, () => {
-            this._textBox.blur();
-            signupState.keyBackedUp = true;
-        });
+        when(() => this.isOk, () => { signupState.keyBackedUp = true; });
     }
+
+    get checkIcon() { return icons.colored('check', null, vars.peerioTeal); }
 
     render() {
         return (
@@ -65,30 +68,27 @@ export default class SignupConfirmBackup extends LoginWizardPage {
                 </View>
                 <View style={{ flex: 0.7, flexGrow: 1, alignItems: 'flex-start' }}>
                     <View style={innerSmall}>
-                        <View style={formStyle}>
+                        <View style={textDescription}>
                             <Text style={titleDark}>{tx('title_confirmTitle')}</Text>
                             <Text style={textNormal}>{tx('title_confirmContent1')} </Text>
                             <Text style={textNormal}>{tx('title_confirmContent2')}</Text>
                             <Text style={textNormal}>{tx('title_confirmTextInput', { sample: this.confirmTextSample })}</Text>
-                            <View style={{ flex: 1, flexGrow: 1, justifyContent: 'center' }}>
-                                <TextBox
-                                    customIcon={this.isOk ? 'check' : null}
-                                    disabled={this.isOk}
-                                    ref={ref => { this._textBox = ref; }}
-                                    lowerCase key="usernameLogin"
-                                    state={this}
-                                    name="confirmText"
-                                    placeholder={this.confirmTextSample}
-                                    {...testLabel('confirmText')}
-                                    hint={this.confirmTextSample} />
-                            </View>
+                        </View>
+                        <View style={{ flexShrink: 1 }}>
+                            <StyledTextInput
+                                customIcon={this.isOk && this.checkIcon}
+                                state={this.confirmTextState}
+                                hint={this.confirmTextSample}
+                                ref={ref => { this._textBox = ref; }}
+                                lowerCase
+                                testID="confirmText" />
                         </View>
                     </View>
                     {this.icon}
                 </View>
                 <View style={[row, { justifyContent: 'space-between' }]}>
                     {this.button('button_back', () => signupState.prev())}
-                    {this.button('button_finish', () => signupState.next(), false, !signupState.keyBackedUp)}
+                    {this.button('button_next', () => signupState.next(), false, !signupState.keyBackedUp)}
                 </View>
                 <ActivityOverlay large visible={signupState.isInProgress} />
             </View>