+++ app/components/wizard/wizard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { observer } from 'mobx-react/native';
 import { View, LayoutAnimation, BackHandler } from 'react-native';
-import { observable, reaction } from 'mobx';
+import { observable, reaction, computed } from 'mobx';
 import SafeComponent from '../shared/safe-component';
 import uiState from '../layout/ui-state';
 
@@ -10,6 +10,11 @@ export default class Wizard extends SafeComponent {
     @observable _index = 0;
     @observable pages = [];
     direction = 1;
+
+    @computed get currentPage() {
+        const { index, pages } = this;
+        return index < pages.length ? this[pages[index]]() : null;
+    }
 
     get index() { return this._index; }
     set index(i) { this._index = i; }
@@ -44,9 +49,10 @@ export default class Wizard extends SafeComponent {
     wizard() {
         // return this.pages.map((k, i) => this._animatedContainer(k, this[k](), i));
         const container = { flexGrow: 1 };
+        const component = this.currentPage;
         return (
             <View style={container}>
-                {this[this.pages[this.index]]()}
+                {component}
             </View>
         );
     }