+++ CHANGELOG.md
@@ -2,6 +2,65 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="3.277.0"></a>
+# [3.277.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.276.0...v3.277.0) (2018-05-03)
+
+
+### Bug Fixes
+
+* correct progress [ch6566] ([11be196](https://github.com/PeerioTechnologies/peerio-mobile/commit/11be196))
+
+
+### Features
+
+* don't display indicators if the visible area is bigger than items ([f3ea4ba](https://github.com/PeerioTechnologies/peerio-mobile/commit/f3ea4ba))
+
+
+
+<a name="3.276.0"></a>
+# [3.276.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.275.0...v3.276.0) (2018-05-02)
+
+
+
+<a name="3.275.0"></a>
+# [3.275.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.274.0...v3.275.0) (2018-05-02)
+
+
+### Bug Fixes
+
+* sending image to self in DM no longer causes error [branch ch4646] ([#179](https://github.com/PeerioTechnologies/peerio-mobile/issues/179)) ([c84cf4b](https://github.com/PeerioTechnologies/peerio-mobile/commit/c84cf4b))
+
+
+### Features
+
+* redesign of file view, inline file download ([#181](https://github.com/PeerioTechnologies/peerio-mobile/issues/181)) ([e5560f6](https://github.com/PeerioTechnologies/peerio-mobile/commit/e5560f6))
+
+
+
+<a name="3.274.0"></a>
+# [3.274.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.273.0...v3.274.0) (2018-04-28)
+
+
+
+<a name="3.273.0"></a>
+# [3.273.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.272.0...v3.273.0) (2018-04-28)
+
+
+### Bug Fixes
+
+*  medcryptor social share url ([#182](https://github.com/PeerioTechnologies/peerio-mobile/issues/182)) ([bd10ee7](https://github.com/PeerioTechnologies/peerio-mobile/commit/bd10ee7))
+* add NSLocationAlwaysUsageDescription ([5fe63e8](https://github.com/PeerioTechnologies/peerio-mobile/commit/5fe63e8))
+* IAP copy ([#178](https://github.com/PeerioTechnologies/peerio-mobile/issues/178)) ([0f9371f](https://github.com/PeerioTechnologies/peerio-mobile/commit/0f9371f))
+* payment per year ([#177](https://github.com/PeerioTechnologies/peerio-mobile/issues/177)) ([99986ba](https://github.com/PeerioTechnologies/peerio-mobile/commit/99986ba))
+
+
+### Features
+
+* add userNotFoundSendInvite error ([#180](https://github.com/PeerioTechnologies/peerio-mobile/issues/180)) ([343e4cc](https://github.com/PeerioTechnologies/peerio-mobile/commit/343e4cc))
+* remove unused plan ([ec604dd](https://github.com/PeerioTechnologies/peerio-mobile/commit/ec604dd))
+
+
+
 <a name="3.272.0"></a>
 # [3.272.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.271.0...v3.272.0) (2018-04-18)